# hw12_13_14_15_16_calendar/internal/api/openapi/v1/types.gen.go
// Package apiv1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package apiv1

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	Code    string                  `json:"code"`
	Details *map[string]interface{} `json:"details"`
	Message string                  `json:"message"`
}

// Event defines model for Event.
type Event struct {
	Description *string   `json:"description"`
	EndTime     time.Time `json:"end_time"`
	Id          int64     `json:"id"`

	// NotifyPeriod Продолжительность уведомления (например, "15m", "1h"). Сервер хранит как time.Duration.
	NotifyPeriod *string   `json:"notify_period"`
	StartTime    time.Time `json:"start_time"`
	Title        string    `json:"title"`
	UserId       int64     `json:"user_id"`
}

// EventCreate defines model for EventCreate.
type EventCreate struct {
	Description  *string   `json:"description"`
	EndTime      time.Time `json:"end_time"`
	NotifyPeriod *string   `json:"notify_period"`
	StartTime    time.Time `json:"start_time"`
	Title        string    `json:"title"`
	UserId       int64     `json:"user_id"`
}

// EventUpdatePut defines model for EventUpdatePut.
type EventUpdatePut = Event

// User defines model for User.
type User struct {
	Email openapi_types.Email `json:"email"`
	Id    int64               `json:"id"`
	Name  string              `json:"name"`
}

// UserCreate defines model for UserCreate.
type UserCreate struct {
	Email openapi_types.Email `json:"email"`
	Name  string              `json:"name"`
}

// UserUpdatePut defines model for UserUpdatePut.
type UserUpdatePut = User

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// InternalError defines model for InternalError.
type InternalError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// GetEventsParams defines parameters for GetEvents.
type GetEventsParams struct {
	// DateFrom Начало интервала (включительно) rfc 3339 формат "YYYY-MM-DDTHH:MM:SSZ"
	DateFrom *time.Time `form:"date_from,omitempty" json:"date_from,omitempty"`

	// DateTo Конец интервала (включительно) rfc 3339 формат "YYYY-MM-DDTHH:MM:SSZ"
	DateTo *time.Time `form:"date_to,omitempty" json:"date_to,omitempty"`

	// UserId Фильтр по владельцу события - id пользователя
	UserId *int64 `form:"user_id,omitempty" json:"user_id,omitempty"`
}

// PostEventsJSONRequestBody defines body for PostEvents for application/json ContentType.
type PostEventsJSONRequestBody = EventCreate

// PutEventsIdJSONRequestBody defines body for PutEventsId for application/json ContentType.
type PutEventsIdJSONRequestBody = EventUpdatePut

// PostUsersJSONRequestBody defines body for PostUsers for application/json ContentType.
type PostUsersJSONRequestBody = UserCreate

// PutUsersIdJSONRequestBody defines body for PutUsersId for application/json ContentType.
type PutUsersIdJSONRequestBody = UserUpdatePut